Rename `unstable_discord_api` to `unstable` and disable it by default (#2789)
This also disables the `#[non_exhaustive]` attribute on `FullEvent` if
`unstable` is enabled as a feature.

It is already expected that any change to Serenity with `unstable` enabled may
or may not break code using this feature and making this enum exhaustive allows
for use cases such as fine-grained audit logging, where new variants should be
a compiler error.

Co-authored-by: Gnome! <david2005thomas@Gmail.com>