Merge `ShardQueuer` functionality into `ShardManager` (#3108)
Untangles the spaghetti of shard management code by removing the
middleman `ShardQueuer` struct, which used a disorienting mess of mpsc
channels to facilitate message passing between the `ShardManager` and
each of its `ShardRunner`s.

The new `ShardManager::run` function takes care of creating/spawning new
runners as needed. This is fine to call without spawning a separate task
(which was being done for `ShardQueuer`) because
`Client::start_connection` previously already waited for the
`ShardQueuer` loop to (essentially) finish before returning. In other
words, the Client would start the `ShardManager` and then block, but the
waiting was just being facilitated through channels, which wasn't
obvious.

The only messages passed across channels now are `ShardRunnerMessage`
(sent from the manager to a given runner), and a new
`ShardManagerMessage` (sent from a runner back to the manager). For
example, the manager can send a message to a runner telling it to
restart, then the runner will shut itself down and ask the manager to
reboot the shard by sending it a message back.

The shard queue now keeps track of if it should dispatch in concurrent
batches or not. Since we only use concurrent startup when first
initializing all shards, and then switch it off once and don't re-enable
it, this should be fine.