Avoid extra Guild removals in Ready handler (#3331)
This code was removing guilds from cache which the current shard is
responsible for, but wasn't received in the unavailable payload.

This code path could only be hit under extreme circumstances, such as
unsuccessful resumes, and since we cannot rely on cache consistency
after an unsuccessful resume anyway this does more bad than good.

To add to the motivation for doing this, out of the libraries checked
(twilight, discord.py, hikari) none do this either.