Add `flags` to `EditWebhookMessage` and `EditInteractionResponse` (#3380)
This commit adds the `flags` field/setter to `EditWebhookMessage` and
`EditInteractionResponse`. It also modifies `EditMessage::suppress_embeds`
because the `IS_COMPONENTS_V2` flag can also be set there now.

This allows upgrading webhook messages and interaction responses to components
V2 in an edit. This appears to be required when deferring interaction responses.